loaders: add local loader to anticipation
