+++ packages/pilot/src/containers/Anticipation/Form/style.css
@@ -39,6 +39,7 @@
 .summaryContainer {
   display: flex;
   flex-direction: column;
+  height: 100%;
 }
 
 .summary {
@@ -51,4 +52,17 @@
 .titleInfo {
   align-items: center;
   display: flex;
+}
+
+.overlay {
+  align-items: center;
+  background-color: var(--color-white);
+  border-radius: 4px;
+  display: block;
+  height: 100%;
+  padding: 0;
+  position: absolute;
+  top: 0;
+  width: 100%;
+  z-index: 5;
 }
