+++ packages/pilot/src/pages/Anticipation/index.js
@@ -7,4 +7,3 @@ export {
   destroyAnticipation,
 } from './actions'
 export { default as epic } from './epic'
-
